+++ project023/manager.cpp
@@ -159,8 +159,8 @@ HRESULT CManager::Init(HINSTANCE hInstance, HWND hWnd, BOOL bWindow)
 			//プレイヤー追従カメラの画面位置設定
 			viewport.X = 0;
 			viewport.Y = 0;
-			viewport.Width = SCREEN_WIDTH * 1.0f;
-			viewport.Height = SCREEN_HEIGHT * 1.0f;
+			viewport.Width = (DWORD)(SCREEN_WIDTH * 1.0f);
+			viewport.Height = (DWORD)(SCREEN_HEIGHT * 1.0f);
 			viewport.MinZ = 0.0f;
 			viewport.MaxZ = 1.0f;
